Temperature_watch
This script is made with the propose to run in background while the computer is running something heavy and not supervised. 
This script should be run as root to be able to shutdown the system any time (This is the defense against overheat).
The invoke should be "sudo temperature_monitor.sh <temperature in Celsius>".