Downloads_Directory_Organizer
It's an code suposed to clean your download directory periodicaly (10 minutes After every startup). 
The code cleans the download sending the files resting in the downloads directory for the Documents, Pictures, Videos, Music, Archive (To compressed archives)... (More details in an arquive called "~/.settings/organizer_rules.txt". To add new rules, just follow the patern (There'll be futher explanation in the future))
To install the script, just untar the the Organizer.tar and execute the install script with super-user privileges.