Amnesic_Directory
It's an simple script to create a directory stored in RAM. The motivation behind this is based on the difficulty to clean data from SSD's (Unlike HDD's, who file erasers tools works to wipe off the data), because the wear leveling technology. So for files who should not left traces the use of ramfs is a handy solution.